refactor: remove remote function endpoints
