+++ packages/sandbox/lib/index.ts
@@ -1,13 +1,13 @@
 export * from './api-keys/sandbox-api-key.service.js';
 export { default as sandboxApiKeyService } from './api-keys/sandbox-api-key.service.js';
+export * from './functions/cli-exit-codes.js';
+export * from './functions/command-builders.js';
+export * from './functions/command-output.js';
+export * from './functions/compiler-client.js';
+export * from './functions/deploy-client.js';
+export * from './functions/dryrun-client.js';
 export * from './functions/dryruns.service.js';
-export * from './remote-function/cli-exit-codes.js';
-export * from './remote-function/command-builders.js';
-export * from './remote-function/command-output.js';
-export * from './remote-function/compiler-client.js';
-export * from './remote-function/deploy-client.js';
-export * from './remote-function/dryrun-client.js';
-export * from './remote-function/helpers.js';
-export * from './remote-function/runtime.js';
-export * from './remote-function/sandbox.js';
-export * from './remote-function/shell.js';
+export * from './functions/helpers.js';
+export * from './functions/runtime.js';
+export * from './functions/sandbox.js';
+export * from './functions/shell.js';
