+++ packages/sandbox/lib/local/deploy-client.ts
@@ -1,18 +1,18 @@
 import { randomUUID } from 'node:crypto';
 
 import { execDockerFileAsync, getExecErrorOutput, isExecTimeoutError, rewriteDockerHostForLocalhost, writeContainerFile } from './docker.js';
-import { getDeployErrorCode } from '../remote-function/cli-exit-codes.js';
-import { buildDeployArgs } from '../remote-function/command-builders.js';
-import { buildIndexTs, getFilePaths } from '../remote-function/compiler-client.js';
-import { RemoteFunctionError } from '../remote-function/helpers.js';
+import { getDeployErrorCode } from '../functions/cli-exit-codes.js';
+import { buildDeployArgs } from '../functions/command-builders.js';
+import { buildIndexTs, getFilePaths } from '../functions/compiler-client.js';
+import { RemoteFunctionError } from '../functions/helpers.js';
 import {
     remoteFunctionDeploySandboxTimeoutMs,
     remoteFunctionDeployTimeoutMs,
     remoteFunctionLocalImage,
     remoteFunctionProjectPath
-} from '../remote-function/runtime.js';
+} from '../functions/runtime.js';
 
-import type { DeployRequest, DeployResult } from '../remote-function/deploy-client.js';
+import type { DeployRequest, DeployResult } from '../functions/deploy-client.js';
 
 export async function invokeLocalDeploy(request: DeployRequest): Promise<DeployResult> {
     const containerName = `nango-deploy-${randomUUID().slice(0, 8)}`;
