+++ packages/server/lib/controllers/functions/validation.ts
@@ -6,33 +6,6 @@ const integrationIdSchema = providerConfigKeySchema.refine((value) => value !== 
     message: 'Integration id cannot be "." or ".."'
 });
 
-const remoteFunctionBaseBodySchema = z
-    .object({
-        integration_id: integrationIdSchema,
-        function_name: syncNameSchema,
-        function_type: z.enum(['action', 'sync']),
-        code: z.string().min(1)
-    })
-    .strict();
-
-export const remoteFunctionCompileBodySchema = remoteFunctionBaseBodySchema;
-
-export const remoteFunctionDeployBodySchema = remoteFunctionBaseBodySchema
-    .extend({
-        allow_destructive: z.boolean().default(false)
-    })
-    .strict();
-
-export const remoteFunctionDryrunBodySchema = remoteFunctionBaseBodySchema
-    .extend({
-        connection_id: connectionIdSchema,
-        input: z.unknown().optional(),
-        metadata: z.record(z.string(), z.unknown()).optional(),
-        checkpoint: z.record(z.string(), z.unknown()).optional(),
-        last_sync_date: z.string().datetime().optional()
-    })
-    .strict();
-
 export const functionCompileBodySchema = z
     .object({
         code: z.string().min(1)
