+++ packages/shared/lib/services/plans/definitions.ts
@@ -29,7 +29,6 @@ export const freePlan: PlanDefinition = {
         can_override_docs_connect_url: false,
         can_customize_connect_ui_theme: false,
         can_disable_connect_ui_watermark: false,
-        remote_functions: false,
         sync_function_runtime: 'lambda',
         action_function_runtime: 'lambda',
         webhook_function_runtime: 'lambda',
@@ -74,7 +73,6 @@ export const starterV1Plan: PlanDefinition = {
         can_override_docs_connect_url: false,
         can_customize_connect_ui_theme: false,
         can_disable_connect_ui_watermark: false,
-        remote_functions: false,
         lambda_tenant_isolation: true
     }
 };
@@ -115,7 +113,6 @@ export const growthV1Plan: PlanDefinition = {
         can_override_docs_connect_url: false,
         can_customize_connect_ui_theme: true,
         can_disable_connect_ui_watermark: true,
-        remote_functions: false,
         lambda_tenant_isolation: true
     }
 };
@@ -183,7 +180,6 @@ export const enterprisePlan: PlanDefinition = {
         can_override_docs_connect_url: false,
         can_customize_connect_ui_theme: true,
         can_disable_connect_ui_watermark: true,
-        remote_functions: false,
         lambda_tenant_isolation: true
     }
 };
@@ -224,7 +220,6 @@ export const starterLegacyPlan: PlanDefinition = {
         can_override_docs_connect_url: false,
         can_customize_connect_ui_theme: false,
         can_disable_connect_ui_watermark: false,
-        remote_functions: false,
         lambda_tenant_isolation: true
     }
 };
@@ -264,7 +259,6 @@ export const scaleLegacyPlan: PlanDefinition = {
         can_override_docs_connect_url: false,
         can_customize_connect_ui_theme: false,
         can_disable_connect_ui_watermark: false,
-        remote_functions: false,
         lambda_tenant_isolation: true
     }
 };
@@ -304,7 +298,6 @@ export const growthLegacyPlan: PlanDefinition = {
         can_override_docs_connect_url: false,
         can_customize_connect_ui_theme: true,
         can_disable_connect_ui_watermark: true,
-        remote_functions: false,
         lambda_tenant_isolation: true
     }
 };
