+++ packages/types/lib/functions/api.ts
@@ -150,63 +150,6 @@ export type PostFunctionDeployment = Endpoint<{
     Success: FunctionDeploySuccess;
 }>;
 
-export interface RemoteFunctionCompileBody {
-    integration_id: string;
-    function_name: string;
-    function_type: RunnableFunctionType;
-    code: string;
-}
-
-export interface RemoteFunctionCompileSuccess extends FunctionCompileSuccess {
-    integration_id: string;
-    function_name: string;
-    function_type: RunnableFunctionType;
-}
-
-export interface RemoteFunctionDryrunBody extends RemoteFunctionCompileBody {
-    connection_id: string;
-    input?: unknown;
-    metadata?: Record<string, unknown> | undefined;
-    checkpoint?: Record<string, unknown> | undefined;
-    last_sync_date?: string | undefined;
-}
-
-export interface RemoteFunctionDryrunSuccess {
-    integration_id: string;
-    function_name: string;
-    function_type: RunnableFunctionType;
-    duration_ms: number;
-    result?: unknown;
-}
-
-export interface RemoteFunctionDeployBody extends RemoteFunctionCompileBody {
-    allow_destructive?: boolean | undefined;
-}
-
-export type PostRemoteFunctionCompile = Endpoint<{
-    Method: 'POST';
-    Path: '/remote-function/compile';
-    Body: RemoteFunctionCompileBody;
-    Error: ApiError<FunctionErrorCode>;
-    Success: RemoteFunctionCompileSuccess;
-}>;
-
-export type PostRemoteFunctionDryrun = Endpoint<{
-    Method: 'POST';
-    Path: '/remote-function/dryrun';
-    Body: RemoteFunctionDryrunBody;
-    Error: ApiError<FunctionErrorCode>;
-    Success: RemoteFunctionDryrunSuccess;
-}>;
-
-export type PostRemoteFunctionDeploy = Endpoint<{
-    Method: 'POST';
-    Path: '/remote-function/deploy';
-    Body: RemoteFunctionDeployBody;
-    Error: ApiError<FunctionErrorCode>;
-    Success: FunctionDeploySuccess;
-}>;
-
 export type GetIntegrationFunctions = Endpoint<{
     Method: 'GET';
     Path: '/api/v1/integrations/:providerConfigKey/functions';
