+++ packages/types/lib/plans/db.ts
@@ -158,12 +158,6 @@ export interface DBPlan extends Timestamps {
     can_disable_connect_ui_watermark: boolean;
 
     /**
-     * Enable or disable remote function compile, dry-run, and deploy endpoints
-     * @default false
-     */
-    remote_functions: boolean;
-
-    /**
      * Sync Function Runtime
      * @default "runner"
      */
